Fix sign of output vector of derivative operator
There is a sign error in the output of the derivative operator in the document. Another PR also concerns this but the proposed fix is not correct, both values need to be fixed.  The other part of the derivation is correct, see output below.

https://github.com/Foundations-of-Computer-Vision/visionbook/pull/49

```
[[-1.  1.  0.  0.  0.]
 [ 0. -1.  1.  0.  0.]
 [ 0.  0. -1.  1.  0.]
 [ 0.  0.  0. -1.  1.]]
Matrix 5* inv A (5x5):
[[-4. -3. -2. -1.]
 [ 1. -3. -2. -1.]
 [ 1.  2. -2. -1.]
 [ 1.  2.  3. -1.]
 [ 1.  2.  3.  4.]]

Signal l (5x1):
[[1.]
 [1.]
 [2.]
 [2.]
 [0.]]

Product r = A * l:
[[ 0.]
 [ 1.]
 [ 0.]
 [-2.]]

Product lhat = pinv(A) * t:
[[-0.2]
 [-0.2]
 [ 0.8]
 [ 0.8]
 [-1.2]]```